Added multiple frames support to the NavigationService library by changing  the page activation pattern.
Added a new sample app with multiple frames being managed by the library.

Updated the single-frame sample app to work with the new library activation patterns.

Updated SuspensionManager to work better with the rest of the library and multiple frames. Ended up deleting a bunch of code in the process. Win win!